requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.7.0
djangorestframework==3.12.4
passlib==1.7.4
python-decouple==3.4
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
